Move domain models to own package
