+++ app/src/main/java/de/elanev/studip/android/app/authorization/data/repository/AuthorizationEntityDataMapper.java
@@ -13,8 +13,8 @@ import javax.inject.Singleton;
 
 import de.elanev.studip.android.app.authorization.data.entity.EndpointEntity;
 import de.elanev.studip.android.app.authorization.data.entity.OAuthCredentialsEntity;
-import de.elanev.studip.android.app.authorization.domain.Endpoint;
-import de.elanev.studip.android.app.authorization.domain.OAuthCredentials;
+import de.elanev.studip.android.app.authorization.domain.model.Endpoint;
+import de.elanev.studip.android.app.authorization.domain.model.OAuthCredentials;
 
 /**
  * @author joern
