+++ app/src/main/java/de/elanev/studip/android/app/authorization/domain/model/OAuthCredentials.java
@@ -1,12 +1,12 @@
 /*
- * Copyright (c) 2016 ELAN e.V.
+ * Copyright (c) 2017 ELAN e.V.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the GNU Public License v3.0
  * which accompanies this distribution, and is available at
  * http://www.gnu.org/licenses/gpl.html
  */
 
-package de.elanev.studip.android.app.authorization.domain;
+package de.elanev.studip.android.app.authorization.domain.model;
 
 /**
  * @author joern
